Consistently use unsigned short for network ports
Network ports are now consistently stored as unsigned shorts, apart from
the options and user interface, where -1 is still used to denote the use
of default ports.

Test-Information:

Unit tests pass on macOS 10.13 and Debian 9
On macOS: tested the UI with various proxy and manual ports, behaviour
as expected.

+++ Slimber/Server.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2016 Isode Limited.
+ * Copyright (c) 2010-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
@@ -39,8 +39,8 @@
 namespace Swift {
 
 Server::Server(
-        int clientConnectionPort,
-        int linkLocalConnectionPort,
+        unsigned short clientConnectionPort,
+        unsigned short linkLocalConnectionPort,
         LinkLocalServiceBrowser* linkLocalServiceBrowser,
         VCardCollection* vCardCollection,
         EventLoop* eventLoop) :
