+++ Swiften/Component/ComponentConnector.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2016 Isode Limited.
+ * Copyright (c) 2010-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
@@ -28,7 +28,7 @@ namespace Swift {
         public:
             typedef std::shared_ptr<ComponentConnector> ref;
 
-            static ComponentConnector::ref create(const std::string& hostname, int port, DomainNameResolver* resolver, ConnectionFactory* connectionFactory, TimerFactory* timerFactory) {
+            static ComponentConnector::ref create(const std::string& hostname, unsigned short port, DomainNameResolver* resolver, ConnectionFactory* connectionFactory, TimerFactory* timerFactory) {
                 return ref(new ComponentConnector(hostname, port, resolver, connectionFactory, timerFactory));
             }
 
@@ -40,7 +40,7 @@ namespace Swift {
             boost::signals2::signal<void (std::shared_ptr<Connection>)> onConnectFinished;
 
         private:
-            ComponentConnector(const std::string& hostname, int port, DomainNameResolver*, ConnectionFactory*, TimerFactory*);
+            ComponentConnector(const std::string& hostname, unsigned short port, DomainNameResolver*, ConnectionFactory*, TimerFactory*);
 
             void handleAddressQueryResult(const std::vector<HostAddress>& address, boost::optional<DomainNameResolveError> error);
             void tryNextAddress();
@@ -53,7 +53,7 @@ namespace Swift {
 
         private:
             std::string hostname;
-            int port;
+            unsigned short port;
             DomainNameResolver* resolver;
             ConnectionFactory* connectionFactory;
             TimerFactory* timerFactory;
